DBZ-8160 Renamed AsyncEngineSignaler to EmbeddedEngineSignaler as it no longer depends on Async engine
