+++ debezium-embedded/src/main/java/io/debezium/embedded/EmbeddedEngineSignaler.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.embedded.async;
+package io.debezium.embedded;
 
 import java.util.List;
 
@@ -11,10 +11,10 @@ import io.debezium.engine.DebeziumEngine;
 import io.debezium.pipeline.signal.SignalRecord;
 import io.debezium.pipeline.signal.channels.process.SignalChannelWriter;
 
-public class AsyncEngineSignaler implements DebeziumEngine.Signaler {
+public class EmbeddedEngineSignaler implements DebeziumEngine.Signaler {
     private final List<SignalChannelWriter> channels;
 
-    public AsyncEngineSignaler(List<SignalChannelWriter> channels) {
+    public EmbeddedEngineSignaler(List<SignalChannelWriter> channels) {
         this.channels = channels;
     }
